Add gobi enable feature

+++ main.py
@@ -9,7 +9,7 @@ load_dotenv()
 
 TOKEN = os.getenv("TOKEN")
 
-intents = discord.Intents.all()  
+intents = discord.Intents.all()
 intents.messages = True
 intents.guilds = True
 
@@ -30,7 +30,7 @@ async def on_ready():
 @bot.event
 async def on_message(message):
     # 語尾dbに書き込んでたら削除
-    if message.channel.name == "語尾db":
+    if message.channel.name == "語尾db" and message.author != bot.user:
         await message.delete()
     if message.author != bot.user:
         print(f"{message.author.display_name}: {message.content}")
@@ -41,12 +41,12 @@ async def on_message(message):
         return
     if message.content.startswith("```"):
         return
-    await get_suffix_channel(message) # 語尾DBチャンネルを取得
+    await replace_suffix(message) # 語尾DBチャンネルを取得
 
     await bot.process_commands(message)
 
 # VCの滞在時間を記録
-@bot.event 
+@bot.event
 async def on_voice_state_update(member, before, after):
     await record_time(member, before, after)
